requirements.txt
python-telegram-bot==20.7
openai==1.3.0
python-dotenv==1.0.0
pydantic==2.4.2
aiohttp==3.9.1